+++ CloudySkies/ModConfig.cs
@@ -7,6 +7,10 @@ public class ModConfig {
 
 	public bool ShowDebugTiming { get; set; } = false;
 
+	public bool RecompileShaders { get; set; } = false;
+
 	public bool ShowWeatherTooltip { get; set; } = true;
+
+	public bool UseCustomGingerIsleArt { get; set; } = true;
 
 }